gut out dirty/clean file from qlibrary_display  and begin developer_mode dev
