+++ qiskit_metal/_gui/widgets/qlibrary_display/proxy_model_qlibrary.py
@@ -39,16 +39,6 @@ class LibraryFileProxyModel(QSortFilterProxyModel):
         # (QComponent files) OR (Directories)
         self.accepted_files__regex = r"(^((?!\.))(?!__init__)(?!_template)(?!__pycache__).*\.py)|(?!__pycache__)(?!base)(^([^.]+)$)"  # pylint: disable=line-too-long
         self.setFilterRegExp(self.accepted_files__regex)
-        self.is_dev_mode = False
-
-    def set_dev_mode(self, ison: bool):
-        """
-        Set dev mode
-        Args:
-            ison: Whether dev mode is on
-
-        """
-        self.is_dev_mode = ison
 
     def filterAcceptsColumn(self, source_column:int, source_parent:QModelIndex) -> bool: #pylint: disable=unused-argument
         """
@@ -60,36 +50,8 @@ class LibraryFileProxyModel(QSortFilterProxyModel):
 
 
         """
-        if (self.is_dev_mode and source_column <= self.sourceModel().REBUILD):
-            return True
         # Won't show Size, Kind, Date Modified, etc. for QFileSystemModel
-        if source_column <= self.sourceModel().REBUILD:
-            return True
+        if source_column > self.sourceModel().FILENAME:
+            return False
 
-        return False
-
-    def data(self,
-             index: QModelIndex,
-             role: int = Qt.DisplayRole) -> typing.Any:
-        """
-        Sets standard size hint for indexes and allows
-        Args:
-            index: Model Index holding data
-            role: DisplayRole being requested of index
-        Returns:
-            Any (QVariant): Data stored under the given role for the item referred to by the index.
-        """
-
-        # allow editable
-        if role == Qt.EditRole:
-            return self.data(index, Qt.DisplayRole)
-
-        if role == Qt.SizeHintRole:
-            return QSize(10, 25)
-
-        # Don't return any data for REBUILD column - Delegate handles that
-        if role == Qt.DisplayRole and index.column() == self.sourceModel(
-        ).REBUILD:
-            return ""
-
-        return super().data(index, role)
+        return True
